refactor: move types to utils
