+++ src/utils/types/types.ts
@@ -1,0 +1,6 @@
+export type TourDateType = {
+  venue: string;
+  location: string;
+  event_date: string;
+  ticket_link: string;
+};
